Fixing MeetingsTest
